divide project into submodules, use separate modules for arguments, commands and core logic
