+++ src/arguments/sanitize.rs
@@ -1,4 +1,4 @@
-use crate::case::string::sanitize;
+use crate::core::case::string::sanitize;
 
 pub fn maybe_sanitize(file_name: String, is_sanitize: &bool) -> String {
     if *is_sanitize {
