+++ src/commands/recase/mod.rs
@@ -1,0 +1,3 @@
+pub mod string;
+pub mod file;
+pub mod files;
